[KARAF-4824] Add Option to bundle:update which doesn't rewrite MANIFEST file

+++ bundle/core/src/main/java/org/apache/karaf/bundle/command/Update.java
@@ -25,6 +25,7 @@ import java.net.URL;
 
 import org.apache.karaf.shell.api.action.Argument;
 import org.apache.karaf.shell.api.action.Command;
+import org.apache.karaf.shell.api.action.Option;
 import org.apache.karaf.shell.api.action.lifecycle.Service;
 import org.apache.karaf.util.bundles.BundleUtils;
 import org.osgi.framework.Bundle;
@@ -37,6 +38,9 @@ public class Update extends BundleCommand {
 
     @Argument(index = 1, name = "location", description = "The bundles update location", required = false, multiValued = false)
     URI location;
+
+    @Option(name = "--raw", description = "Do not update the bundles's Bundle-UpdateLocation manifest header")
+    boolean raw;
 
     protected Object doExecute(Bundle bundle) throws Exception {
         if (location != null) {
@@ -54,11 +58,15 @@ public class Update extends BundleCommand {
 
     private void update(Bundle bundle, URL location) throws IOException, BundleException {
         try (InputStream is = location.openStream()) {
-            File file = BundleUtils.fixBundleWithUpdateLocation(is, location.toString());
-            try (FileInputStream fis = new FileInputStream(file)) {
-                bundle.update(fis);
+            if (raw) {
+                bundle.update(is);
+            } else {
+                File file = BundleUtils.fixBundleWithUpdateLocation(is, location.toString());
+                try (FileInputStream fis = new FileInputStream(file)) {
+                    bundle.update(fis);
+                }
+                file.delete();
             }
-            file.delete();
         }
     }
